refactor: errors -> common
