+++ candle/common/frontend.py
@@ -3,6 +3,7 @@ from flask_login import current_user
 
 from candle import db
 from candle.models import UserTimetable
+from flask_wtf.csrf import CSRFError
 
 common = Blueprint('common', __name__, template_folder="templates",
                 static_folder='static',
@@ -26,3 +27,23 @@ def home():
         return redirect(url_for('my_timetable.show_timetable', id_=ut.id_) )
     else:  # user is logged out, show welcome-info:
         return render_template('timetable/timetable.html', title='Rozvrh', show_welcome=True)
+
+
+@common.app_errorhandler(404)
+def error_404(error):
+    return render_template('errors/404.html'), 404
+
+
+@common.app_errorhandler(403)
+def error_403(error):
+    return render_template('errors/403.html'), 403
+
+
+@common.app_errorhandler(500)
+def error_500(error):
+    return render_template('errors/500.html'), 500
+
+
+@common.app_errorhandler(CSRFError)
+def csrf_error(reason):
+    return render_template('errors/csrf_error.html', reason=reason.description), 400
