+++ candle/common/templates/errors/500.html
